Expose a jni wrapper for rocks-db cloud.

+++ cloud/cloud_file_system.cc
@@ -81,8 +81,8 @@ void CloudFileSystemOptions::Dump(Logger* log) const {
   Header(log, "COptions.delete_cloud_invisible_files_on_open: %d",
          delete_cloud_invisible_files_on_open);
   if (cloud_file_deletion_delay) {
-    Header(log, "          COptions.cloud_file_deletion_delay: %ld",
-           cloud_file_deletion_delay->count());
+    Header(log, "          COptions.cloud_file_deletion_delay: %lld",
+           static_cast<long long>(cloud_file_deletion_delay->count()));
   }
 }
